refactor: using double for price

+++ ViewModels/Car/AddCarViewModel.cs
@@ -5,5 +5,5 @@ public record AddCarViewModel
     [BsonElement("Brand")] public string VehicleBrand { get; init; } = string.Empty!;
 
     public string Name { get; init; } = string.Empty!;
-    public int Price { get; init; }
+    public double Price { get; init; }
 }
