+++ ViewModels/Car/UpdateCarViewModel.cs
@@ -8,5 +8,5 @@ public record UpdateCarViewModel
 
     [BsonElement("Brand")] public string VehicleBrand { get; init; } = string.Empty!;
     public string Name { get; init; } = string.Empty!;
-    public int Price { get; init; }
+    public double Price { get; init; }
 }
